+++ spec/lightstreamer/stream_connection_header_spec.rb
@@ -11,16 +11,39 @@ describe Lightstreamer::StreamConnectionHeader do
     expect(header['ControlAddress']).to eq('abc')
   end
 
-  it 'processes a failure header in the expected format' do
+  it 'processes an error header' do
     expect(header.process_header_line('ERROR')).to be true
-    expect(header.process_header_line('10')).to be true
-    expect(header.process_header_line('Error message')).to be false
-    expect(header.error.code).to eq(10)
-    expect(header.error.error).to eq('Error message')
+    expect(header.process_header_line('1')).to be true
+    expect(header.process_header_line('')).to be true
+
+    expect(header.error).to be_a(Lightstreamer::AuthenticationError)
   end
 
-  it 'handles an unexpected format' do
-    expect(header.process_header_line('Unexpected error format')).to be false
-    expect(header.error.error).to eq('Unexpected error format')
+  it 'processes an end header' do
+    expect(header.process_header_line('END')).to be true
+
+    expect(header.error).to be_a(Lightstreamer::SessionEndError)
+  end
+
+  it 'processes an end header with a cause code' do
+    expect(header.process_header_line('END')).to be true
+    expect(header.process_header_line('31')).to be true
+
+    expect(header.error).to be_a(Lightstreamer::SessionEndError)
+    expect(header.error.cause_code).to eq(31)
+  end
+
+  it 'processes a sync error' do
+    expect(header.process_header_line('SYNC ERROR')).to be false
+
+    expect(header.error).to be_a(Lightstreamer::SyncError)
+  end
+
+  it 'processes an unrecognized error' do
+    expect(header.process_header_line('Unknown')).to be true
+    expect(header.process_header_line('data')).to be true
+
+    expect(header.error).to be_a(Lightstreamer::Error)
+    expect(header.error.message).to eq('Unknown data')
   end
 end